Moved code to ensure exception messages do not mess up test XML output to XML generator. And do full XML escaping instead of just replacing < and >.
Original reason to process message by IlseH:
Added code to remove characters that interfere with Surefire report
watching. For SLIM this is no issue because an earlier assertion error
is raised in case of a failure that has no invalid characters and
therefore can be processed by Surefire.
However for FIT, somehow only these last assertions are raised and the
&lt; and &gt; characters cause the TEST-null.xml not to be processed by
Surefire so any failure logged here was not seen.