Reaction to suggested changes.
I have performed all the requested changes.
As to whether the fixture tests are more useful when they exist as a Junit file or a FitNesse wiki table, I plead neutrality. I agree it may make more sense in this project to place the tests mostly or even only in FitNesse tables.