Fixes to let FileFixture(Test) work on Windows and closing of handles
The FileFixtureTest was not working on Windows because a linux path
seperator was used. When correcting the test, it seemed some handles
were not closed (could not remove file) so fixed this in FileFixture. In
addition we seem to be able to use
org.apache.commons.io.FileUtils.copyFile, instead of copying the
contents ourselves.