Use latest travis image (but keep sudo). --- From: support@travis-ci.com Subject: Re: Chrome startup from selenium fails Date: 16 January 2018 at 13:20:42 CET
Hi Fried,

Thanks for writing in and sorry for the troubles.

I've been looking into the build you shared with us, and I see that this one ran in our `sudo: false` environment.

Unfortunately, this is a side-effect of the work we're doing to mitigate the Spectre/Meltdown exploit, which required us to perform maintenance on our container-based infrastructure. One of the side effects of this is that it's caused issues with Chrome running on that infrastructure, probably due to the underlying kernel update.

At the moment, we recommend routing these builds to our `sudo: required` infrastructure, which seems to have worked correctly for you. In your case, I believe that the `group: deprecated-2017Q4` is not really having an effect in fixing the issue, but the `sudo: required` option is.

Alternatively, it's also possible to keep your builds in `sudo: false` and disable the sandbox by adding the --no-sandbox flag in the command you use to run Chrome.

Does this help? If you're curious, you can find out more information in this public issue https://github.com/travis-ci/travis-ci/issues/8836#issuecomment-357441064

I hope this helps! Please let me know if you have any other questions or concerns.